requirements.txt
peewee==3.17.9
flask==3.1.0
apiflask==2.1.0  # Для документации API и валидации
PyJWT==2.9.0     # Для работы с JWT-токенами
passlib==1.7.4   # Для хэширования паролей пользователей